+++ .prettierrc.json
@@ -1,0 +1,4 @@
+{
+    "singleQuote": true,
+    "semi": false
+}
