+++ index.html
@@ -2,12 +2,12 @@
 <html lang="en">
     <head>
         <meta charset="UTF-8" />
-        <link rel="icon" type="image/svg+xml" href="/R.png" />
+        <link rel="icon" type="image/svg+xml" href="/client/assets/R.png" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>RedTetris</title>
     </head>
     <body>
         <div id="root"></div>
-        <script type="module" src="/src/main.jsx"></script>
+        <script type="module" src="/client/main.jsx"></script>
     </body>
 </html>
