+++ package.json
@@ -4,7 +4,7 @@
     "version": "0.0.0",
     "type": "module",
     "scripts": {
-        "dev": "vite",
+        "dev": "APP_PORT=3004 nodemon server/server.js",
         "build": "vite build",
         "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
         "preview": "vite preview"
@@ -12,8 +12,11 @@
     "dependencies": {
         "@fortawesome/free-solid-svg-icons": "^6.5.1",
         "@fortawesome/react-fontawesome": "^0.2.0",
+        "express": "^4.19.2",
         "react": "^18.2.0",
-        "react-dom": "^18.2.0"
+        "react-dom": "^18.2.0",
+        "socket.io": "^4.7.5",
+        "socket.io-client": "^4.7.5"
     },
     "devDependencies": {
         "@types/react": "^18.2.66",
@@ -23,6 +26,7 @@
         "eslint-plugin-react": "^7.34.1",
         "eslint-plugin-react-hooks": "^4.6.0",
         "eslint-plugin-react-refresh": "^0.4.6",
+        "nodemon": "^3.1.0",
         "vite": "^5.2.0"
     }
 }
