+++ server/game.js
@@ -1,0 +1,1 @@
+export class Game {}
